requirements.txt
Django
mysqlclient
requests
pillow
djangorestframework